# backend/src/main/java/com/watchtogether/backend/feedback/FeedbackReportListResponse.java
package com.watchtogether.backend.feedback;

import java.time.Instant;
import java.util.List;

public record FeedbackReportListResponse(
        Instant listedAt,
        int count,
        List<FeedbackReportSummary> reports) {}