# contracts/openapi.yaml
openapi: 3.1.0
jsonSchemaDialect: https://json-schema.org/draft/2020-12/schema

info:
  title: Watch Together REST API
  version: 1.0.0
  description: |
    Контракт product backend. Эндпоинты с `x-implementation-status: planned`
    фиксируют границу для P2 и не считаются реализованными до соответствующего тикета.

servers:
  - url: /
    description: Same-origin API через development или production reverse proxy

tags:
  - name: System
    description: Системные endpoints
  - name: Rooms
    description: Жизненный цикл приватной комнаты
  - name: Feedback
    description: Обратная связь closed beta и operator triage
  - name: Telemetry
    description: Privacy-safe client telemetry для beta-метрик

paths:
  /api/v1/health:
    get:
      tags: [System]
      operationId: getHealth
      summary: Проверить доступность backend
      x-implementation-status: implemented
      responses:
        "200":
          description: Backend доступен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /api/v1/version:
    get:
      tags: [System]
      operationId: getVersion
      summary: Получить версию backend
      x-implementation-status: implemented
      responses:
        "200":
          description: Версия backend
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionResponse"

  /api/v1/feedback:
    post:
      tags: [Feedback]
      operationId: submitFeedback
      summary: Принять обратную связь по beta-сессии
      description: |
        Endpoint принимает короткий session feedback и optional privacy-safe browser/network
        metadata. Он не принимает медиафайлы, chat history, room secrets или LiveKit tokens.
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackRequest"
      responses:
        "202":
          description: Отзыв принят и получил correlation-friendly receipt
          headers:
            Cache-Control:
              description: Ответ содержит диагностический receipt и не должен кэшироваться
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/feedback/reports:
    get:
      tags: [Feedback]
      operationId: listFeedbackReports
      summary: Получить последние beta feedback reports
      description: |
        Operator endpoint для регулярного triage. Доступен только при настроенном
        `FEEDBACK_ADMIN_TOKEN` и корректном header `X-Feedback-Admin-Token`.
      x-implementation-status: implemented
      security:
        - feedbackAdminToken: []
      parameters:
        - $ref: "#/components/parameters/FeedbackLimit"
      responses:
        "200":
          description: Последние reports в компактном виде
          headers:
            Cache-Control:
              description: Operator данные не кэшируются
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReportListResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/feedback/reports/export:
    get:
      tags: [Feedback]
      operationId: exportFeedbackReports
      summary: Выгрузить последние beta feedback reports
      description: |
        Operator export для ручного разбора beta-прогона. Возвращает полные reports
        в пределах `FEEDBACK_EXPORT_LIMIT`; медиа, chat history, room secrets и LiveKit
        tokens здесь не хранятся и не возвращаются.
      x-implementation-status: implemented
      security:
        - feedbackAdminToken: []
      parameters:
        - $ref: "#/components/parameters/FeedbackLimit"
      responses:
        "200":
          description: Последние reports в полном виде
          headers:
            Cache-Control:
              description: Operator данные не кэшируются
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReportExportResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/feedback/reports/{feedbackId}:
    parameters:
      - $ref: "#/components/parameters/FeedbackId"
    get:
      tags: [Feedback]
      operationId: getFeedbackReport
      summary: Получить один beta feedback report
      x-implementation-status: implemented
      security:
        - feedbackAdminToken: []
      responses:
        "200":
          description: Полный report
          headers:
            Cache-Control:
              description: Operator данные не кэшируются
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReport"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Feedback]
      operationId: triageFeedbackReport
      summary: Обновить triage поля beta feedback report
      x-implementation-status: implemented
      security:
        - feedbackAdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackTriageRequest"
      responses:
        "200":
          description: Report после triage update
          headers:
            Cache-Control:
              description: Operator данные не кэшируются
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/telemetry:
    post:
      tags: [Telemetry]
      operationId: submitTelemetry
      summary: Принять батч privacy-safe client telemetry
      description: |
        Endpoint принимает батч client telemetry событий (first frame, publish/playback
        ошибки, quality summary) для агрегированных beta-метрик. Он не принимает медиафайлы,
        chat history, room secrets, LiveKit tokens или identity данные. Каждое событие даёт
        только aggregate counter и low-cardinality tag; roomId, role и detail используются
        лишь для log-корреляции и никогда не становятся метрик-тегами.
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryRequest"
      responses:
        "202":
          description: Батч принят и получил correlation-friendly receipt
          headers:
            Cache-Control:
              description: Ответ содержит диагностический receipt и не должен кэшироваться
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/rooms:
    post:
      tags: [Rooms]
      operationId: createRoom
      summary: Создать приватную комнату
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "201":
          description: Комната создана, host session установлена
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookie"
            Location:
              description: API-адрес созданной комнаты
              schema:
                type: string
                pattern: ^/api/v1/rooms/
            Cache-Control:
              description: Ответ содержит credentials и не должен кэшироваться
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoomResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/rooms/{roomId}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [Rooms]
      operationId: getRoom
      summary: Восстановить room session и authoritative snapshot
      x-implementation-status: implemented
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Текущий participant и snapshot комнаты для активной browser session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRoomResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/RoomUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/rooms/{roomId}/join:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [Rooms]
      operationId: joinRoom
      summary: Присоединить гостя к комнате
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomRequest"
      responses:
        "200":
          description: Гость присоединён или его session identity восстановлена
          headers:
            Set-Cookie:
              $ref: "#/components/headers/SessionCookie"
            Cache-Control:
              description: Ответ содержит session credential и не должен кэшироваться
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRoomResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/RoomUnavailable"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/rooms/{roomId}/leave:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [Rooms]
      operationId: leaveRoom
      summary: Покинуть комнату от имени guest
      x-implementation-status: implemented
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Guest participant покинул комнату
          headers:
            Cache-Control:
              description: Ответ не должен кэшироваться
              schema:
                type: string
                const: no-store
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RoomUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/rooms/{roomId}/livekit-token:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [Rooms]
      operationId: mintLiveKitToken
      summary: Выдать LiveKit token для текущего participant
      x-implementation-status: implemented
      security:
        - sessionCookie: []
      responses:
        "200":
          description: LiveKit token с grants, соответствующими роли текущего participant
          headers:
            Cache-Control:
              description: Ответ содержит credential и не должен кэшироваться
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiveKitTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/RoomUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/rooms/{roomId}/close:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [Rooms]
      operationId: closeRoom
      summary: Закрыть комнату от имени host
      x-implementation-status: implemented
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/HostSecret"
      responses:
        "204":
          description: Комната закрыта или уже была закрыта
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/RoomUnavailable"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: wt_session
      description: |
        Opaque session credential. Production cookie: HttpOnly, Secure, SameSite=Strict.
    feedbackAdminToken:
      type: apiKey
      in: header
      name: X-Feedback-Admin-Token
      description: Operator token from `FEEDBACK_ADMIN_TOKEN`; disabled when env is empty.

  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        $ref: "./schemas/common.schema.json#/$defs/RoomId"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Повтор того же запроса возвращает исходный результат; другой payload даёт conflict.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[!-~]+$
    HostSecret:
      name: X-Host-Secret
      in: header
      required: true
      description: Opaque host secret. Никогда не помещается в invite URL или логи.
      schema:
        type: string
        minLength: 43
        maxLength: 128
        pattern: ^[A-Za-z0-9_-]+$
    FeedbackId:
      name: feedbackId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FeedbackLimit:
      name: limit
      in: query
      required: false
      description: Ограничение размера выборки; backend clamp-ит значение к FEEDBACK_EXPORT_LIMIT.
      schema:
        type: integer
        minimum: 1
        maximum: 1000

  headers:
    SessionCookie:
      description: HttpOnly session cookie; Secure обязателен вне локальной разработки.
      schema:
        type: string

  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status, checkedAt]
      properties:
        status:
          type: string
          const: UP
        checkedAt:
          type: string
          format: date-time

    VersionResponse:
      type: object
      additionalProperties: false
      required: [name, version, buildTime, apiVersion]
      properties:
        name:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
        buildTime:
          type: string
          minLength: 1
        apiVersion:
          type: string
          const: v1

    FeedbackOutcome:
      type: string
      enum: [WORKED, ISSUE, BLOCKED]

    FeedbackReason:
      type: string
      enum:
        - SUCCESS
        - CONNECTION
        - AUDIO_VIDEO
        - FILE
        - VOICE
        - SYNC
        - CHAT
        - ROOM_ACCESS
        - PERFORMANCE
        - OTHER

    FeedbackClientMetadata:
      type: object
      additionalProperties: false
      properties:
        userAgent:
          type: string
          maxLength: 512
        language:
          type: string
          maxLength: 32
        platform:
          type: string
          maxLength: 64
        viewportWidth:
          type: integer
          minimum: 0
          maximum: 10000
        viewportHeight:
          type: integer
          minimum: 0
          maximum: 10000
        devicePixelRatio:
          type: number
          minimum: 0
          maximum: 10
        networkEffectiveType:
          type: string
          maxLength: 32
        networkDownlinkMbps:
          type: number
          minimum: 0
          maximum: 10000
        networkRttMs:
          type: integer
          minimum: 0
          maximum: 60000
        networkSaveData:
          type: boolean
        roomStatus:
          type: string
          maxLength: 32
        roomConnectionStatus:
          type: string
          maxLength: 32
        liveKitStatus:
          type: string
          maxLength: 32
        qualityStatus:
          type: string
          maxLength: 32
        participantCount:
          type: integer
          minimum: 0
          maximum: 4

    FeedbackRequest:
      type: object
      additionalProperties: false
      required: [outcome, reason]
      properties:
        outcome:
          $ref: "#/components/schemas/FeedbackOutcome"
        reason:
          $ref: "#/components/schemas/FeedbackReason"
        message:
          type: string
          maxLength: 2000
        roomId:
          $ref: "./schemas/common.schema.json#/$defs/RoomId"
        participantRole:
          type: string
          enum: [HOST, GUEST]
        relatedCorrelationId:
          type: string
          format: uuid
        metadata:
          $ref: "#/components/schemas/FeedbackClientMetadata"

    FeedbackResponse:
      type: object
      additionalProperties: false
      required: [feedbackId, correlationId, receivedAt]
      properties:
        feedbackId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        receivedAt:
          type: string
          format: date-time

    FeedbackTriageStatus:
      type: string
      enum: [NEW, REVIEWING, RESOLVED, IGNORED]

    FeedbackSeverity:
      type: string
      enum: [UNSET, LOW, MEDIUM, HIGH, BLOCKER]

    FeedbackTriageRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/FeedbackTriageStatus"
        severity:
          $ref: "#/components/schemas/FeedbackSeverity"
        assignee:
          type: string
          maxLength: 120
        note:
          type: string
          maxLength: 1000

    FeedbackReportSummary:
      type: object
      additionalProperties: false
      required:
        - feedbackId
        - correlationId
        - receivedAt
        - outcome
        - reason
        - triageStatus
        - severity
      properties:
        feedbackId:
          type: string
          format: uuid
        correlationId:
          type: string
        receivedAt:
          type: string
          format: date-time
        outcome:
          $ref: "#/components/schemas/FeedbackOutcome"
        reason:
          $ref: "#/components/schemas/FeedbackReason"
        roomId:
          $ref: "./schemas/common.schema.json#/$defs/RoomId"
        participantRole:
          type: string
          enum: [HOST, GUEST]
        relatedCorrelationId:
          type: string
          format: uuid
        triageStatus:
          $ref: "#/components/schemas/FeedbackTriageStatus"
        severity:
          $ref: "#/components/schemas/FeedbackSeverity"
        assignee:
          type: string
          maxLength: 120
        messagePreview:
          type: string
          maxLength: 160

    FeedbackReport:
      type: object
      additionalProperties: false
      required:
        - feedbackId
        - correlationId
        - receivedAt
        - outcome
        - reason
        - triageStatus
        - severity
      properties:
        feedbackId:
          type: string
          format: uuid
        correlationId:
          type: string
        receivedAt:
          type: string
          format: date-time
        outcome:
          $ref: "#/components/schemas/FeedbackOutcome"
        reason:
          $ref: "#/components/schemas/FeedbackReason"
        message:
          type: string
          maxLength: 2000
        roomId:
          $ref: "./schemas/common.schema.json#/$defs/RoomId"
        participantRole:
          type: string
          enum: [HOST, GUEST]
        relatedCorrelationId:
          type: string
          format: uuid
        metadata:
          $ref: "#/components/schemas/FeedbackClientMetadata"
        triageStatus:
          $ref: "#/components/schemas/FeedbackTriageStatus"
        severity:
          $ref: "#/components/schemas/FeedbackSeverity"
        assignee:
          type: string
          maxLength: 120
        triageNote:
          type: string
          maxLength: 1000
        triagedAt:
          type: string
          format: date-time

    FeedbackReportListResponse:
      type: object
      additionalProperties: false
      required: [listedAt, count, reports]
      properties:
        listedAt:
          type: string
          format: date-time
        count:
          type: integer
          minimum: 0
        reports:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackReportSummary"

    FeedbackReportExportResponse:
      type: object
      additionalProperties: false
      required: [exportedAt, count, reports]
      properties:
        exportedAt:
          type: string
          format: date-time
        count:
          type: integer
          minimum: 0
        reports:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackReport"

    TelemetryEventType:
      type: string
      enum:
        - FIRST_FRAME
        - PLAYBACK_ERROR
        - PUBLISH_START
        - PUBLISH_FAILURE
        - QUALITY_SUMMARY

    TelemetryQualityStatus:
      type: string
      enum: [GOOD, WARNING, POOR, LOST, UNKNOWN]

    TelemetryEvent:
      type: object
      additionalProperties: false
      required: [type]
      properties:
        type:
          $ref: "#/components/schemas/TelemetryEventType"
        roomId:
          $ref: "./schemas/common.schema.json#/$defs/RoomId"
        role:
          type: string
          enum: [HOST, GUEST]
        qualityStatus:
          $ref: "#/components/schemas/TelemetryQualityStatus"
        detail:
          type: string
          maxLength: 200

    TelemetryRequest:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/TelemetryEvent"

    TelemetryResponse:
      type: object
      additionalProperties: false
      required: [telemetryId, correlationId, receivedAt, accepted]
      properties:
        telemetryId:
          type: string
          format: uuid
        correlationId:
          type: string
          format: uuid
        receivedAt:
          type: string
          format: date-time
        accepted:
          type: integer
          minimum: 0

    RoomSnapshot:
      $ref: "./schemas/common.schema.json#/$defs/RoomSnapshot"

    Participant:
      $ref: "./schemas/common.schema.json#/$defs/Participant"

    CreateRoomRequest:
      type: object
      additionalProperties: false
      required: [hostDisplayName]
      properties:
        hostDisplayName:
          $ref: "./schemas/common.schema.json#/$defs/DisplayName"

    CreateRoomResponse:
      type: object
      additionalProperties: false
      required: [room, hostSecret, invitePath]
      properties:
        room:
          $ref: "#/components/schemas/RoomSnapshot"
        hostSecret:
          type: string
          readOnly: true
          minLength: 43
          maxLength: 128
          pattern: ^[A-Za-z0-9_-]+$
        invitePath:
          type: string
          pattern: ^/rooms/[A-Za-z0-9_-]{22}$
          description: Guest URL path без host secret.

    JoinRoomRequest:
      type: object
      additionalProperties: false
      required: [displayName]
      properties:
        displayName:
          $ref: "./schemas/common.schema.json#/$defs/DisplayName"

    JoinRoomResponse:
      type: object
      additionalProperties: false
      required: [participant, room]
      properties:
        participant:
          $ref: "#/components/schemas/Participant"
        room:
          $ref: "#/components/schemas/RoomSnapshot"

    GetRoomResponse:
      type: object
      additionalProperties: false
      required: [participant, room]
      properties:
        participant:
          $ref: "#/components/schemas/Participant"
        room:
          $ref: "#/components/schemas/RoomSnapshot"

    LiveKitTokenResponse:
      type: object
      additionalProperties: false
      required:
        - token
        - liveKitUrl
        - roomName
        - participantId
        - participantIdentity
        - role
        - canPublish
        - canPublishData
        - expiresAt
      properties:
        token:
          type: string
          minLength: 20
          description: Signed LiveKit JWT. Нельзя логировать или кэшировать.
        liveKitUrl:
          type: string
          pattern: ^wss?://
          description: Browser-facing LiveKit URL.
        roomName:
          $ref: "./schemas/common.schema.json#/$defs/RoomId"
        participantId:
          type: string
          format: uuid
        participantIdentity:
          type: string
          minLength: 1
          maxLength: 128
        role:
          type: string
          enum: [HOST, GUEST]
        canPublish:
          type: boolean
        canPublishData:
          type: boolean
        expiresAt:
          type: string
          format: date-time

    ProblemDetails:
      $ref: "./schemas/problem-details.schema.json"

  responses:
    BadRequest:
      description: Некорректный JSON или формат запроса
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Unauthorized:
      description: Session credential отсутствует или недействительна
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: Участник не имеет права выполнить операцию
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    RoomUnavailable:
      description: Комната недоступна; ответ не раскрывает, существовала ли она
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: Ресурс не найден или уже удалён по retention policy
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Conflict:
      description: Конфликт состояния, idempotency key или лимита участников
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationFailed:
      description: Запрос не прошёл предметную валидацию
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    RateLimited:
      description: Превышен допустимый темп запросов
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    InternalError:
      description: Внутренняя ошибка без раскрытия реализации
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"